update ElastiCache impl instead of new class
